# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for a Lorcana tracker.",
        "title": "Lorcana Tracker API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/card": {
            "get": {
                "description": "Get all cards",
                "produces": [
                    "application/json"
                ],
                "summary": "Get all cards",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/card.Card"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/card/{id}": {
            "get": {
                "description": "Get a card by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a card by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Card ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/card.Card"
                        }
                    },
                    "404": {
                        "description": "no card found with the given ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/card_in_collection": {
            "get": {
                "description": "Get all cards in collection",
                "produces": [
                    "application/json"
                ],
                "summary": "Get all cards in collection",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/card.InCollection"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a card in collection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a card in collection",
                "parameters": [
                    {
                        "description": "Card in collection",
                        "name": "cardInCollection",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/card.InCollection"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/card.InCollection"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/card_in_collection/{id}": {
            "get": {
                "description": "Get a card in collection by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a card in collection by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Card ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/card.InCollection"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/card_price/{id}": {
            "get": {
                "description": "Get a card price by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a card price by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Card ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/card.Price"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/set": {
            "get": {
                "description": "Get all sets",
                "produces": [
                    "application/json"
                ],
                "summary": "Get all sets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/card.Set"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "card.Card": {
            "type": "object",
            "properties": {
                "artist": {
                    "type": "string"
                },
                "body_text": {
                    "type": "string"
                },
                "card_num": {
                    "type": "integer"
                },
                "color": {
                    "type": "string"
                },
                "cost": {
                    "type": "integer"
                },
                "flavor_text": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "inkable": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "rarity": {
                    "type": "string"
                },
                "set_id": {
                    "type": "string"
                },
                "set_name": {
                    "type": "string"
                },
                "set_num": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "card.InCollection": {
            "type": "object",
            "required": [
                "card_id",
                "owned_foil_copies",
                "owned_normal_copies",
                "wish_list"
            ],
            "properties": {
                "card_id": {
                    "type": "string"
                },
                "owned_foil_copies": {
                    "type": "integer"
                },
                "owned_normal_copies": {
                    "type": "integer"
                },
                "wish_list": {
                    "type": "boolean"
                }
            }
        },
        "card.Price": {
            "type": "object",
            "required": [
                "card_id"
            ],
            "properties": {
                "card_id": {
                    "type": "string"
                },
                "market_price_in_euro": {
                    "type": "integer"
                },
                "market_price_link": {
                    "type": "string"
                }
            }
        },
        "card.Set": {
            "type": "object",
            "properties": {
                "set_id": {
                    "type": "string"
                },
                "set_name": {
                    "type": "string"
                },
                "set_num": {
                    "type": "integer"
                }
            }
        }
    }
}